Mejoras en la implementación AddCustomerActivity.kt Pendiente cambios textField en el LoginActivity

+++ app/src/main/java/com/itesthida/techserviceapp/data/database/entities/Equipment.kt
@@ -1,14 +1,14 @@
 package com.itesthida.techserviceapp.data.database.entities
 
 data class Equipment(
-    val id: Int,
+    val id: Long,
     val customer: Customer,
     val equipmentType: EquipmentType,
     val serialNumber: String
 ) {
     companion object{
         // Constante para el id por defecto
-        const val DEFAULT_ID = -1
+        const val DEFAULT_ID = -1L
         // Constantes para los datos de la tabla, nombre de tabla y columnas
         const val TABLE_NAME = "EQUIPMENTS"
         const val COLUMN_NAME_ID = "ID"
