+++ app/src/main/java/com/itesthida/techserviceapp/data/database/entities/EquipmentType.kt
@@ -1,14 +1,14 @@
 package com.itesthida.techserviceapp.data.database.entities
 
 data class EquipmentType(
-    val id: Int,
+    val id: Long,
     val name: String,
     val brand: String,
     val model: String
 ) {
     companion object{
         // Constante para el id por defecto
-        const val DEFAULT_ID = -1
+        const val DEFAULT_ID = -1L
         // Constantes para los datos de la tabla, nombre de tabla y columnas
         const val TABLE_NAME = "EQUIPMENT_TYPES"
         const val COLUMN_NAME_ID = "ID"
