+++ app/src/main/java/com/itesthida/techserviceapp/data/database/entities/State.kt
@@ -1,12 +1,12 @@
 package com.itesthida.techserviceapp.data.database.entities
 
 data class State(
-    val id: Int,
+    val id: Long,
     val stateName: String
 ) {
     companion object{
         // Constante para el di por defecto
-        const val DEFAULT_ID = -1
+        const val DEFAULT_ID = -1L
         // Constantes para los datos de la tabla, nombre de tabla y columnas
         const val TABLE_NAME = "STATES"
         const val COLUMN_NAME_ID = "ID"
