+++ app/src/main/java/com/itesthida/techserviceapp/data/database/entities/Technician.kt
@@ -1,12 +1,8 @@
 package com.itesthida.techserviceapp.data.database.entities
 
-//import androidx.room.Entity
-//import androidx.room.PrimaryKey
 
-//@Entity(tableName = "Technicians")
 data class Technician (
-    //@PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val name: String,
     val lastName: String,
     val email: String,
@@ -14,7 +10,7 @@ data class Technician (
 ){
     companion object{
         // Constante para el id por defecto
-        const val DEFAULT_ID = -1
+        const val DEFAULT_ID = -1L
         // Constantes para los datos de la tabla, nombre de tabla y columnas
         const val TABLE_NAME = "TECHNICIANS"
         const val COLUMN_NAME_ID = "ID"
